requirements.txt
pika
openai
python-dotenv
numpy